requirements.txt
numpy==1.23.1
opencv-python==4.5.3.56
Pillow==9.2.0
